+++ src/crawlee/basic_crawler/_basic_crawler.py
@@ -55,10 +55,10 @@ if TYPE_CHECKING:
     from crawlee.events._event_manager import EventManager
     from crawlee.http_clients import BaseHttpClient, HttpResponse
     from crawlee.proxy_configuration import ProxyConfiguration, ProxyInfo
+    from crawlee.request_sources._request_provider import RequestProvider
     from crawlee.sessions import Session
     from crawlee.statistics import FinalStatistics, StatisticsState
     from crawlee.storages._dataset import ExportDataCsvKwargs, ExportDataJsonKwargs, GetDataKwargs, PushDataKwargs
-    from crawlee.storages._request_provider import RequestProvider
 
 TCrawlingContext = TypeVar('TCrawlingContext', bound=BasicCrawlingContext, default=BasicCrawlingContext)
 ErrorHandler = Callable[[TCrawlingContext, Exception], Awaitable[Union[Request, None]]]
