+++ src/crawlee/request_sources/__init__.py
@@ -1,0 +1,11 @@
+from ._request_source import RequestSource  # noqa: I001 - out of order to prevent circular import
+from ._request_list import RequestList
+from ._request_provider import RequestProvider
+from ._request_source_tandem import RequestSourceTandem
+
+__all__ = [
+    'RequestList',
+    'RequestProvider',
+    'RequestSource',
+    'RequestSourceTandem',
+]
