+++ src/crawlee/request_sources/_request_list.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 from typing_extensions import override
 
 from crawlee._utils.docs import docs_group
-from crawlee.storages._request_source import RequestSource
+from crawlee.request_sources import RequestSource
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
