+++ src/crawlee/request_sources/_request_provider.py
@@ -5,7 +5,7 @@ from datetime import timedelta
 from typing import TYPE_CHECKING
 
 from crawlee._utils.docs import docs_group
-from crawlee.storages._request_source import RequestSource
+from crawlee.request_sources import RequestSource
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
