+++ src/crawlee/storages/__init__.py
@@ -1,17 +1,9 @@
 from ._dataset import Dataset
 from ._key_value_store import KeyValueStore
-from ._request_list import RequestList
-from ._request_provider import RequestProvider
 from ._request_queue import RequestQueue
-from ._request_source import RequestSource
-from ._request_source_tandem import RequestSourceTandem
 
 __all__ = [
     'Dataset',
     'KeyValueStore',
-    'RequestList',
-    'RequestProvider',
     'RequestQueue',
-    'RequestSource',
-    'RequestSourceTandem',
 ]
